Factor out xAPI utilities
